Exposes dedupe via dedupeUtils
