+++ packages/plugin-essentials/sources/commands/dedupe.ts
@@ -12,104 +12,12 @@
  * (which also provides a safe-guard in case virtual descriptors ever make their way into the dedupe algorithm).
  */
 
-import {BaseCommand}                                                                                                        from '@yarnpkg/cli';
-import {Configuration, Project, ResolveOptions, ThrowReport, Cache, StreamReport, Resolver, miscUtils, Descriptor, Package} from '@yarnpkg/core';
-import {formatUtils, structUtils, IdentHash, LocatorHash, MessageName, Report, Fetcher, FetchOptions}                       from '@yarnpkg/core';
-import {Command}                                                                                                            from 'clipanion';
-import micromatch                                                                                                           from 'micromatch';
-import * as yup                                                                                                             from 'yup';
+import {BaseCommand}                                 from '@yarnpkg/cli';
+import {Configuration, Project, Cache, StreamReport} from '@yarnpkg/core';
+import {Command}                                     from 'clipanion';
+import * as yup                                      from 'yup';
 
-export type DedupePromise = Promise<{
-  descriptor: Descriptor,
-  currentPackage: Package,
-  updatedPackage: Package,
-} | null>;
-
-export type DedupeAlgorithm = (project: Project, patterns: Array<string>, opts: {
-  resolver: Resolver,
-  resolveOptions: ResolveOptions,
-  fetcher: Fetcher,
-  fetchOptions: FetchOptions,
-}) => Promise<Array<DedupePromise>>;
-
-export enum Strategy {
-  /**
-   * This strategy dedupes a locator to the best candidate already installed in the project.
-   *
-   * Because of this, it's guaranteed that:
-   * - it never takes more than a single pass to dedupe all dependencies
-   * - dependencies are never downgraded
-   */
-  HIGHEST = `highest`,
-}
-
-const acceptedStrategies = new Set(Object.values(Strategy));
-
-export const DEDUPE_ALGORITHMS: Record<Strategy, DedupeAlgorithm> = {
-  highest: async (project, patterns, {resolver, fetcher, resolveOptions, fetchOptions}) => {
-    const locatorsByIdent = new Map<IdentHash, Set<LocatorHash>>();
-    for (const [descriptorHash, locatorHash] of project.storedResolutions) {
-      const descriptor = project.storedDescriptors.get(descriptorHash);
-      if (typeof descriptor === `undefined`)
-        throw new Error(`Assertion failed: The descriptor (${descriptorHash}) should have been registered`);
-
-      miscUtils.getSetWithDefault(locatorsByIdent, descriptor.identHash).add(locatorHash);
-    }
-
-    return Array.from(project.storedDescriptors.values(), async descriptor => {
-      if (patterns.length && !micromatch.isMatch(structUtils.stringifyIdent(descriptor), patterns))
-        return null;
-
-      const currentResolution = project.storedResolutions.get(descriptor.descriptorHash);
-      if (typeof currentResolution === `undefined`)
-        throw new Error(`Assertion failed: The resolution (${descriptor.descriptorHash}) should have been registered`);
-
-      // We only care about resolutions that are stored in the lockfile
-      // (we shouldn't accidentally try deduping virtual packages)
-      const currentPackage = project.originalPackages.get(currentResolution);
-      if (typeof currentPackage === `undefined`)
-        return null;
-
-      // No need to try deduping packages that are not persisted,
-      // they will be resolved again anyways
-      if (!resolver.shouldPersistResolution(currentPackage, resolveOptions))
-        return null;
-
-      const locators = locatorsByIdent.get(descriptor.identHash);
-      if (typeof locators === `undefined`)
-        throw new Error(`Assertion failed: The resolutions (${descriptor.identHash}) should have been registered`);
-
-      // No need to choose when there's only one possibility
-      if (locators.size === 1)
-        return null;
-
-      const references = [...locators].map(locatorHash => {
-        const pkg = project.originalPackages.get(locatorHash);
-        if (typeof pkg === `undefined`)
-          throw new Error(`Assertion failed: The package (${locatorHash}) should have been registered`);
-
-        return pkg.reference;
-      });
-
-      const candidates = await resolver.getSatisfying(descriptor, references, resolveOptions);
-
-      const bestCandidate = candidates?.[0];
-      if (typeof bestCandidate === `undefined`)
-        return null;
-
-      const updatedResolution = bestCandidate.locatorHash;
-
-      const updatedPackage = project.originalPackages.get(updatedResolution);
-      if (typeof updatedPackage === `undefined`)
-        throw new Error(`Assertion failed: The package (${updatedResolution}) should have been registered`);
-
-      if (updatedResolution === currentResolution)
-        return null;
-
-      return {descriptor, currentPackage, updatedPackage};
-    });
-  },
-};
+import * as dedupeUtils                              from '../dedupeUtils';
 
 // eslint-disable-next-line arca/no-default-export
 export default class DedupeCommand extends BaseCommand {
@@ -117,7 +25,7 @@ export default class DedupeCommand extends BaseCommand {
   patterns: Array<string> = [];
 
   @Command.String(`-s,--strategy`, {description: `The strategy to use when deduping dependencies`})
-  strategy: Strategy = Strategy.HIGHEST;
+  strategy: dedupeUtils.Strategy = dedupeUtils.Strategy.HIGHEST;
 
   @Command.Boolean(`-c,--check`, {description: `Exit with exit code 1 when duplicates are found, without persisting the dependency tree`})
   check: boolean = false;
@@ -129,9 +37,9 @@ export default class DedupeCommand extends BaseCommand {
     strategy: yup.string().test({
       name: `strategy`,
       message: `\${path} must be one of \${strategies}`,
-      params: {strategies: [...acceptedStrategies].join(`, `)},
+      params: {strategies: [...dedupeUtils.acceptedStrategies].join(`, `)},
       test: (strategy: string) => {
-        return acceptedStrategies.has(strategy as Strategy);
+        return dedupeUtils.acceptedStrategies.has(strategy as dedupeUtils.Strategy);
       },
     }),
   });
@@ -192,7 +100,7 @@ export default class DedupeCommand extends BaseCommand {
       stdout: this.context.stdout,
       json: this.json,
     }, async report => {
-      dedupedPackageCount = await dedupe({project, strategy: this.strategy, patterns: this.patterns, cache, report});
+      dedupedPackageCount = await dedupeUtils.dedupe(project, {strategy: this.strategy, patterns: this.patterns, cache, report});
     });
 
     if (dedupeReport.hasErrors())
@@ -212,99 +120,4 @@ export default class DedupeCommand extends BaseCommand {
       return installReport.exitCode();
     }
   }
-}
-
-export type DedupeSpec = {
-  strategy: Strategy,
-  project: Project,
-  patterns: Array<string>,
-  cache: Cache,
-  report: Report,
-};
-
-export async function dedupe({strategy, project, patterns, cache, report}: DedupeSpec) {
-  const {configuration} = project;
-  const throwReport = new ThrowReport();
-
-  const resolver = configuration.makeResolver();
-  const fetcher = configuration.makeFetcher();
-
-  const fetchOptions: FetchOptions = {
-    cache,
-    checksums: project.storedChecksums,
-    fetcher,
-    project,
-    report: throwReport,
-    skipIntegrityCheck: true,
-  };
-  const resolveOptions: ResolveOptions = {
-    project,
-    resolver,
-    report: throwReport,
-    fetchOptions,
-  };
-
-  return await report.startTimerPromise(`Deduplication step`, async () => {
-    const algorithm = DEDUPE_ALGORITHMS[strategy];
-    const dedupePromises = await algorithm(project, patterns, {resolver, resolveOptions, fetcher, fetchOptions});
-
-    const progress = StreamReport.progressViaCounter(dedupePromises.length);
-    report.reportProgress(progress);
-
-    let dedupedPackageCount = 0;
-
-    await Promise.all(
-      dedupePromises.map(dedupePromise =>
-        dedupePromise
-          .then(dedupe => {
-            if (dedupe === null)
-              return;
-
-            dedupedPackageCount++;
-
-            const {descriptor, currentPackage, updatedPackage} = dedupe;
-
-            report.reportInfo(
-              MessageName.UNNAMED,
-              `${
-                structUtils.prettyDescriptor(configuration, descriptor)
-              } can be deduped from ${
-                structUtils.prettyLocator(configuration, currentPackage)
-              } to ${
-                structUtils.prettyLocator(configuration, updatedPackage)
-              }`
-            );
-
-            report.reportJson({
-              descriptor: structUtils.stringifyDescriptor(descriptor),
-              currentResolution: structUtils.stringifyLocator(currentPackage),
-              updatedResolution: structUtils.stringifyLocator(updatedPackage),
-            });
-
-            project.storedResolutions.set(descriptor.descriptorHash, updatedPackage.locatorHash);
-          })
-          .finally(() => progress.tick())
-      )
-    );
-
-    let packages: string;
-    switch (dedupedPackageCount) {
-      case 0: {
-        packages = `No packages`;
-      } break;
-
-      case 1: {
-        packages = `One package`;
-      } break;
-
-      default: {
-        packages = `${dedupedPackageCount} packages`;
-      }
-    }
-
-    const prettyStrategy = formatUtils.pretty(configuration, strategy, formatUtils.Type.CODE);
-    report.reportInfo(MessageName.UNNAMED, `${packages} can be deduped using the ${prettyStrategy} strategy`);
-
-    return dedupedPackageCount;
-  });
 }
